updated handling of 500 error. reset connection on network timeout
Fixes ZPS-1354

If we get a 500 error trying to stop the get-counter cmdlet, let's
try it twice more.  continue processing so that we can delete the
remote shell.

Also, if we get a ConnectError, this could be a network timeout that
occurs on a connection closed by other side, but not by us.  This way
we won't keep trying to send requests on a dead connection.