Fix 'kerberos' referenced before assignment
DeferredSemaphore.run(fn, *args, **kwargs) has the method signature it
does because it defers the creation of the fn(*args, **kwargs) deferred
until the semaphore can be acquired. My change to wrap a timeout around
all functions run under the semaphore in e5ca422 broke this by causing
fn(*args, **kwargs) to be executed before the semaphore could be
acquired.

This change restores the execution of fn(*args, **kwargs) to after the
semaphore is acquired, and thus fixes the resulting 'kerberos'
referenced before assignment problems it caused in the Windows ZenPack.

Refs ZPS-1795.