Keep number of open shells as low as possible
This update will allow us to use a single shell per device.  On a service
startup, like zenpython, we'll probably end up with a few shells.  After
running for a while, all clients will start to use the longest running
shell owned by a specific user.  After a daemon restart, if the shell has
not timed out, we can reconnect to it and continue as if nothing happened.
On Session logout, remove the shell for the device ip.  On the next query,
retrieve the oldest running session.  This will be good for services like
zenmodeler that do not run very often.

This also takes into account older versions of windows zp or other zenpacks
that are not using this version of txwinrm.