Fix so that we will use a single shell for short commands
Revert reusing a single shell for lifetime of collection

Fixes ZPS-2132

A Windows command session becomes invalid after deleting the remote shell.
What we will do is always create a new shell and send x number of commands
to that shell.  When we are closing the connection, we will delete the shell.
We'll also keep a separate session for enumerations.  These are quick hitters
and don't need to be slowed by waiting for a shell command to complete.

Just before deleting a shell, we'll make sure that there are no clients waiting
to use it.  Then we'll remove it from our dictionary so that we don't create a new
client while waiting for the current shell to be deleted.

Long commands will create a shell, then delete it when done.

Revert single shell usage for single/long commands

remove unused options