Add connection timeout option, possible memory leak fix
Fixes ZPS-2829

Added a connection timeout so that we can return quicker from a failed connection
This will allow for fewer tasks to be in a running state.  connections should only take a few seconds at most
Pool connections seem to keep growing in memory.  Destroy agent and connection pool when closing a connection
Also allow for starting an immediate close connection