updates for dealing with closing connections. also kerberos
Fixes ZPS-2394, ZPS-2963

during modeling, we were obtaining spn tickets in parallel when trying to query wmi and powershell.  this is creating two tickets in the cache.  modeling was still working, but we really should only do this once.
the agent was not being reallocated for basic auth for a long running command so we need to make sure we get a new agent
also made it so that we consistently close a connection when attempting a retry or when we are finished with a connection
a better way to abort a close connection is to see if the deferred has been cancelled