return an OK response, kerb context fixes
Fixes ZPS-5510

when we were retrying a request we were raising an error if it turned out to be ok. now we'll return the response data. other fixes around an expiring ticket and readability.  improved testing if session was connected.  created a test in collect.py to prove the expiration handling.  the test will defer until there's less than 60s left in the ticket lifetime, then allow the handler to wait for the ticket to expire and create a new connection