try to complete stop so we delete shells
fixes ZPS-5598

we should continue to try and complete the stop process so we don't leave orphaned processes on the target.

fix decrypt integrity check, no remote shell errors

ensure that we use the correct kerberos context to decrypt received data.  gssclient could be overwritten in the time between receiving the body and the time we try to decrypt it.  send in shell_cmd pair into receive() so we do not see no remote shell errors while stopping a command