account for possible missing authentication header from windows
Fixes ZPS-5770

if windows server does not return the WWW-Authenticate header in the response, we should handle this.  this is some sort of error with windows responses, maybe due to out of date Windows Management Framework.  there are many bugs in the old WMF, and this could be one of them.