simplify winrm connections
fixes ZPS-3757, ZPS-5822, ZPS-5797

This update will no longer use Session Management. There were too many moving parts and points of failure.  We will tie the request and connection close together so that we are not destroying a connection while trying to create a one to use.  each instance of a client will maintain it's own connection to a windows device.  fixed the delay of an expiring kerberos ticket so that it actually "sleeps" while waiting for it to expire.  for each request, we will wait for the previous request to finish, then reuse the connection.  if there is no connection, we will create a new one.  also added "operation_timeout" to the command line so that we can test adjusting the operation timeout of winrm/winrs on the command line.  changed the connect timeout of the agent so that it doesn't sit around waiting for several minutes.  allow the agent to use the default cached connection timeout also