add ability to auto renew kerberos tickets
Fixes ZPS-5797

this will give the ability to automatically renew the kerberos user ticket so that we do not see an issue with attempting to use an expired context.  we will initiate a deferred that will fire 5 minutes (by default) before a ticket expires.  the first connection will add the one and only deferred per user and will call 'kinit -r user@DOMAIN'.  also found that using 'klist -l' may truncate the full user name.  we'll use 'klist -A' instead so that we can get the full user names for all caches.  added a simple test function in collect to make sure that user_in_klist is performing as expected.