be sure to save new connection when retrying request
fixes ZPS-5918

network timeout errors in the log indicate we could be trying to reuse a bad connection.  we need to be sure we're saving off a new connection when it's created and destroy the bad connection.  RequestErors are ok to keep the same connection as they should still be valid.  removed unused imports